feat: add migration

+++ src/Domain/Power/PowerEntity.cs
@@ -7,7 +7,7 @@ namespace Domain.Power
     public class PowerEntity : BaseAuditableEntity<ushort>
     {
         public ushort CardId { get; set; } = default;
-        public uint FeatureId { get; set; } = default;
+        public ushort FeatureId { get; set; } = default;
         public uint Value { get; set; } = default;
         public CardEntity Card { get; set; } = null!;
         public FeatureEntity Feature { get; set; } = null!;
